feat: Refactor MedicationHistoryScreen
Refactored the MedicationHistoryScreen to match the layout and components of the TodaySchedulesScreen.

- Updated the MedicationHistoryViewModel to support grouping by month or week and to efficiently fetch data.
- Replaced the existing layout with a Scaffold and LargeTopAppBar.
- Created a new FilterControls composable with a dropdown for date filtering and a toggle for sorting.
- Replaced the MedicationHistoryListItem with a new HistoryScheduleItem composable, a modified version of TodayScheduleItem.
- Implemented grouping and list display using a LazyColumn.
- Added an EmptyState composable.
- Fixed a bug in the click handler for the HistoryScheduleItem.
- Fixed a performance issue in the MedicationHistoryViewModel by removing N+1 queries.
- Added missing string resources for multiple languages.
- Corrected compilation errors related to incorrect data model properties and missing resources.
